refactor(Error): result関数をsafeExecute関数にリファクタリング
