+++ package/main/src/Error/index.ts
@@ -1,1 +1,1 @@
-export * from "./result";
+export * from "./safeExecute";
