+++ package/main/src/Error/safeExecute.ts
@@ -1,0 +1,23 @@
+interface SuccessType<V> {
+  type: "success";
+  value: V;
+}
+
+interface ErrorType<E> {
+  type: "err";
+  error: E;
+}
+
+export type Result<V, E> = SuccessType<V> | ErrorType<E>;
+const errorFunction = <E>(error: E): ErrorType<E> => ({ type: "err", error });
+const successFunction = <V>(value: V): SuccessType<V> => ({
+  type: "success",
+  value,
+});
+export const safeExecute = <V, E = Error>(callback: () => V): Result<V, E> => {
+  try {
+    return successFunction<V>(callback());
+  } catch (error) {
+    return errorFunction<E>(error as E);
+  }
+};
