+++ package/main/src/tests/Array/shuffle2DArray.test.ts
@@ -1,19 +1,6 @@
 import { shuffle2DArray } from "@/Array/shuffle2DArray";
 
 describe("shuffle2DArray関数のテスト", () => {
-  it("2次元配列全体がシャッフルされること", () => {
-    const array = [
-      [1, 2],
-      [3, 4],
-      [5, 6],
-    ];
-    const shuffledArray = shuffle2DArray(array);
-    // 全体の配列がシャッフルされているか確認
-    expect(shuffledArray.map((subArray) => subArray.sort())).not.toEqual(
-      array.map((subArray) => subArray.sort()),
-    );
-  });
-
   it("空の2次元配列がそのまま返されること", () => {
     const array: number[][] = [];
     const shuffledArray = shuffle2DArray(array);
